+++ src/app/aplication-state.ts
@@ -1,5 +1,4 @@
-import { IInvoiceDetailInitialState } from './invoice/invoice-detail/ngrx/invoice-detail.reducer';
-import { IInvoiceInitialState } from './invoice/invoice-list/ngrx/invoice-list.reducer';
+import { IInvoiceInitialState } from './invoice/ngrx/invoice.reducer';
 import { IProductsInitialState } from './products/ngrx/product.reducer';
 import { IUsersInitialState } from './users/ngrx/user.reducer';
 
@@ -7,5 +6,4 @@ export interface IApplicationState {
   userState: IUsersInitialState;
   productState: IProductsInitialState;
   invoiceState: IInvoiceInitialState;
-  invoiceDetailState: IInvoiceDetailInitialState;
 }
