+++ src/app/invoice/invoice-list/invoice-list.component.ts
@@ -3,13 +3,13 @@ import { ActivatedRoute, Params, Router } from '@angular/router';
 import { Store } from '@ngrx/store';
 import { Observable, Subscription } from 'rxjs';
 import { IApplicationState } from '../../aplication-state';
-import { IInvoice } from '../invoice-invoiceDetail.interface';
-import { loadInvoices } from './ngrx/invoice-list.actions';
+import { IInvoice } from '../invoice.interface';
+import { loadInvoices } from '../ngrx/invoice.actions';
 import {
   loadInvoicesPending,
   selectInvoices,
   error,
-} from './ngrx/invoice-list.selectors';
+} from '../ngrx/invoice.selectors';
 import { MatDialog } from '@angular/material/dialog';
 import { ErrorMessage } from 'src/app/shared/error-message';
