+++ src/app/invoice/invoice.interface.ts
@@ -2,9 +2,6 @@ export interface IInvoice {
   invoiceId: number;
   date: Date;
   customerName: string;
-}
-
-export interface IInvoiceDetail {
   invoiceDetailId: number;
   name: string;
   description: string;
