+++ src/app/invoice/ngrx/invoice.effects.ts
@@ -2,8 +2,8 @@ import { Injectable } from '@angular/core';
 import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { of } from 'rxjs';
 import { map, mergeMap, catchError } from 'rxjs/operators';
-import { loadInvoices } from './invoice-list.actions';
-import { InvoiceService } from '../../invoice.service';
+import { loadInvoices, loadInvoiceDetail } from './invoice.actions';
+import { InvoiceService } from '../invoice.service';
 
 @Injectable()
 export class InvoiceEffects {
@@ -14,6 +14,19 @@ export class InvoiceEffects {
         this.invoiceService.fetchInvoicesPaginated(action.page).pipe(
           map((invoice) => loadInvoices.success({ invoices: invoice })),
           catchError((error) => of(loadInvoices.failure({ error })))
+        )
+      )
+    )
+  );
+  loadInvoiceDetail$ = createEffect(() =>
+    this.actions$.pipe(
+      ofType(loadInvoiceDetail.begin),
+      mergeMap((action) =>
+        this.invoiceService.fetchInvoiceDetail(action.id).pipe(
+          map((invoiceDetail) =>
+            loadInvoiceDetail.success({ invoiceDetail: invoiceDetail })
+          ),
+          catchError((error) => of(loadInvoiceDetail.failure({ error })))
         )
       )
     )
