+++ src/app/invoice/ngrx/invoice.selectors.ts
@@ -3,16 +3,26 @@ import {
   IInvoiceInitialState,
   invoiceFeatureKey,
   selectAll,
-} from './invoice-list.reducer';
+  selectIds,
+} from './invoice.reducer';
 
 export const selectInvoiceState =
   createFeatureSelector<IInvoiceInitialState>(invoiceFeatureKey);
 
 export const selectInvoices = createSelector(selectInvoiceState, selectAll);
+export const selectInvoiceDetail = createSelector(
+  selectInvoiceState,
+  selectIds
+);
 
 export const loadInvoicesPending = createSelector(
   selectInvoiceState,
   (state: IInvoiceInitialState) => state.loadInvoicesPending
+);
+
+export const loadInvoiceDetailPending = createSelector(
+  selectInvoiceState,
+  (state: IInvoiceInitialState) => state.loadInvoiceDetailPending
 );
 export const error = createSelector(
   selectInvoiceState,
